--- src/pages/Chats/Matching/styles.tsx
@@ -1,10 +1,0 @@
-import { styled } from 'styled-components';
-
-import { StyledText } from '@components/Text/StyledText';
-
-export const ReqeustInfo = styled(StyledText)`
-	display: flex;
-	align-items: center;
-	padding: 0.5rem 1.25rem;
-	text-align: left;
-`;
